+++ Novel-Nest/Controllers/CategoryController.cs
@@ -51,30 +51,38 @@ namespace Novel_Nest.Controllers
             return View(model);
         }
 
-        [HttpPost]
-        public async Task<IActionResult> DeleteCategory(int Id)
-        {
-            try
-            {
-                var success = await _categoryService.DeleteCategoryAsync(Id);
-                if (success)
-                {
-                    return RedirectToAction("Category");
-                }
-                else
-                {
-                    ViewBag.ErrorMessage = "Failed to delete category.";
-                    return View("ErrorView");
-                }
-            }
-            catch (Exception ex)
-            {
-                ViewBag.ErrorMessage = "Failed to delete category. " + ex.Message;
-                return View("ErrorView");
-            }
-        }
+		[HttpPost]
+		public async Task<IActionResult> DeleteCategory(int Id)
+		{
+			try
+			{
+				bool isCategoryInUse = await _categoryService.IsCategoryInUseAsync(Id);
+				if (isCategoryInUse)
+				{
+					TempData["ErrorMessage"] = "Cannot delete category because it is in use.";
+					return RedirectToAction("Category");
+				}
 
-        [HttpPost]
+				var success = await _categoryService.DeleteCategoryAsync(Id);
+				if (success)
+				{
+					TempData["Message"] = "Category successfully deleted.";
+					return RedirectToAction("Category");
+				}
+				else
+				{
+					TempData["ErrorMessage"] = "Failed to delete category.";
+					return RedirectToAction("Category");
+				}
+			}
+			catch (Exception ex)
+			{
+				TempData["ErrorMessage"] = "Failed to delete category. " + ex.Message;
+				return RedirectToAction("Category");
+			}
+		}
+
+		[HttpPost]
         public async Task<IActionResult> EditCategoryAsync(CategoryDTO category)
         {
             try
